refactor: movies 상태 및 useEffect 코드를 App.jsx로 이동
